+++ app/controllers/categories_controller.rb
@@ -11,9 +11,8 @@ class CategoriesController < ApplicationController
       @latest_items = @top_items[1..3]
     end
     @categories = @site_categories
-    private_headers
     respond_to do |format|
-      format.html
+      format.html { headers_with_timeout(600) }
     end
   end
 
@@ -47,25 +46,20 @@ class CategoriesController < ApplicationController
       @last_mofified = @last_published
       @etag = Digest::MD5.hexdigest(@items.map{|t| t.id}.to_s)
     end
-    
+
     respond_to do |format|
       format.html {
-        headers_with_timeout(600, 'public')
+        headers_with_timeout(600)
       }
       format.atom {
         headers['Etag'] = @etag
-        headers['Cache-Control'] = 'public, max-age=1200'
-        headers['Last-Modified'] = @last_published.httpdate if @last_published
+        headers_with_timeout(900)
         render partial: "/shared/items", layout: false 
       }
       format.rss {
-        headers['Etag'] = @etag
-        headers['Cache-Control'] = 'public, max-age=1200'
-        headers['Last-Modified'] = @last_published.httpdate if @last_published
+        headers_with_timeout(900)
         render partial: "/shared/items", layout: false
       }
-      format.json { render json: @items }
-      format.xml { render @items }
     end
   end